Add a setting for network polling interval
The polling process may take many seconds to complete if a large filesystem
is connected over a relatively slow link.  If the amount of time required
to do a single pass is longer than the polling update interval then the CPU
load may remain pegged at 100% because XBMC starts a new update as soon as
the previous one has finished.

I have added a new setting called POLLING_INTERVAL which can be used to
configure the amount of time (in seconds) between successive polls.
This setting only applies to the XBMCVFS based polling methods.